Store userdata in 64-bit compatible way

+++ Code/Commando/dlgconfigaudiotab.cpp
@@ -444,7 +444,7 @@ DlgConfigAudioTabClass::Configure_Driver_List (void)
 			//
 			int item_index = list_ctrl->Insert_Entry (index, wide_driver_name);
 			if (item_index >= 0) {
-				list_ctrl->Set_Entry_Data (item_index, 0, (DWORD)driver_info->driver);
+				list_ctrl->Set_Entry_Data (item_index, 0, (uintptr_t)driver_info->driver);
 
 				//
 				//	Select this entry if its the default
